+++ binless/src/gam.cpp
@@ -74,11 +74,13 @@ void GeneralizedAdditiveModel::optimize(const Eigen::VectorXd& y, const Eigen::V
 
 
 
-Eigen::SparseMatrix<double> first_order_difference_matrix(unsigned K) {
-  if (K<=1) return Eigen::SparseMatrix<double>();
+Eigen::SparseMatrix<double> decreasing_constraint(unsigned K, unsigned Kfree) {
+  if (K <= 1 || Kfree >= K) return Eigen::SparseMatrix<double>();
+  //build first order difference matrix
   Eigen::MatrixXi D = - Eigen::MatrixXi::Identity(K-1,K);
   D.diagonal(1).setConstant(1);
-  return D.sparseView().cast<double>();
+  //return only last rows
+  return - D.bottomRows(K-Kfree).sparseView().cast<double>();
 }
 
 Eigen::SparseMatrix<double> second_order_difference_matrix(unsigned K) {