+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 for versions 0.x of binless, minor releases might break backwards compatibility.
 
 ## [Unreleased]
+### Added
+- First part of decay in fast binless is not forced to decrease. Use `free_diag`
+ argument to decide (default 10kb)
+
 ### Changed
 - Exposures are now fit separately
 - Use `nobs` column in fast binless